refactor: reduce code smells
