+++ packages/aksara-ui-core/src/components/form/components/InputRadio/InputRadioLabel.tsx
@@ -1,31 +1,8 @@
-import * as React from 'react';
-import styled from 'styled-components';
-import { Box, BoxProps } from '../../../../foundations';
+import { InputCheckboxLabelProps, InputCheckboxLabel } from '../InputCheckbox';
 
-export interface InputRadioLabelProps extends BoxProps, React.LabelHTMLAttributes<HTMLLabelElement> {
-  className?: string;
-  style?: React.CSSProperties;
-}
+export type InputRadioLabelProps = InputCheckboxLabelProps;
 
-const LabelBase = styled(Box)`
-  cursor: pointer;
-`;
-
-const InputRadioLabel: React.FC<InputRadioLabelProps> = ({ className, style, children, ...rest }) => {
-  return (
-    <LabelBase
-      as="label"
-      display="inline-flex"
-      flexDirection="row"
-      alignItems="center"
-      className={className}
-      style={style}
-      {...rest}
-    >
-      {children}
-    </LabelBase>
-  );
-};
+const InputRadioLabel = InputCheckboxLabel;
 
 InputRadioLabel.displayName = 'InputRadioLabel';
